Autoformat everything with EmmyLuaCodeStyle
This one seems a lil bit more reliable than LuaFormatter and is
conveniently built in to the SumNeko Lua Language Server addon.

This does have the ability to give a decent approximation of Garry's
preferred coding style, but I have decided to turn that off partially
out of spite but mostly because nothing else looks like that.