Resolve all the LLS errors in the database test file
Unfortunately `luaassert` has a wildly complex and non-deterministic
assertion chain system, and the current types provided in the community
repo do not represent a fraction of what is actually available.

I'm going to make a PR to fix this at some point but until then, I've
had to adjust the assertions to match the defined ones.

I've also had to throw a bunch of `disable-next-line` calls around to
make the tests that check type safety not error for being unsafe.